+++ persistence/controllers/user_controller.js
@@ -5,14 +5,21 @@ var Rating = require("../models/Rating.js");
 var Favorite = require("../models/Favorite.js");
 
 const logger = require("../helpers/loggers.js");
+const checkApiKey = require("../helpers/check_apiKey.js");
+const ObjectId = require("mongodb").ObjectID;
 
-router.post("/add", async function (req, res) {
+// Create a new user
+router.post("/", async function (req, res) {
+  if (!checkApiKey(req.headers.authorization)) {
+    logger.error("Unauthorized.");
+    return res.status(401).json({ success: false, message: "Unauthorized." });
+  }
   const newUser = new User({
     username: req.body.username,
     karma: 0,
     role: "user",
     id: req.body.id,
-    joinDate: req.body.joinDate,
+    joinDate: Date.now(),
   });
 
   await User.findOne(
@@ -50,7 +57,12 @@ router.post("/add", async function (req, res) {
   );
 });
 
-router.post("/delete/", async function (req, res) {
+// Delete a user
+router.delete("/", async function (req, res) {
+  if (!checkApiKey(req.headers.authorization)) {
+    logger.error("Unauthorized.");
+    return res.status(401).json({ success: false, message: "Unauthorized." });
+  }
   logger.info(`Request to delete user with ID: ${req.body.id}`);
   await Rating.deleteMany({ userId: req.body.id }).exec(function (
     err,
@@ -78,12 +90,12 @@ router.post("/delete/", async function (req, res) {
     }
   });
 
-  await User.findOneAndDelete({ id: req.body.id }).exec(function (err, user) {
+  await User.deleteOne({ _id: req.body.id }).exec(function (err, user) {
     if (err) {
       logger.error(`Something went wrong ! ${err}`);
       res.status(400).json({ success: false, msg: err });
     } else if (!user) {
-      logger.error(`User not found in database.`);
+      logger.error(`${ObjectId(req.body.id)} not found in database.`);
       res
         .status(404)
         .json({ success: false, msg: "User not found in database." });
